feature(velas-install): 'versions' command, 'init' from github
