+++ install/src/build_env.rs
@@ -1,2 +1,1 @@
 pub const TARGET: &str = env!("TARGET");
-pub const BUILD_SECONDS_SINCE_UNIX_EPOCH: &str = env!("BUILD_SECONDS_SINCE_UNIX_EPOCH");
